ft(trip request): update endpoints

+++ src/controllers/tripRequestController.js
@@ -38,12 +38,12 @@ export const getSingleTripRequest = async (req, res) => {
         const requestId = req.params.id;
 
         if (req.user.role == 'requester') {
-            const trips = await tripRequests.findOne({ where: { requesterId: userId, } })
+            const trip = await tripRequests.findOne({ where: { id: requestId, requesterId: userId, } })
 
-            if (!trips) {
+            if (!trip) {
                 return res.status(404).json({ status: 'fail', message: `there are no Trip Requests assigned to user with id=  ${userId}` })
             } else {
-                return res.status(200).json({ status: 'success', data: trips })
+                return res.status(200).json({ status: 'success', data: trip });
             }
 
         } else if (req.user.role == 'travel admin') {
@@ -57,7 +57,6 @@ export const getSingleTripRequest = async (req, res) => {
         } else {
             return res.status(403).json({ status: 'fail', message: 'you are not allowed to retrieve a trip request' })
         }
-
 
     } catch (error) {
         res.status(500).json({ status: 'error', message: 'error while getting trip request' })
@@ -99,17 +98,21 @@ export const updateTripRequest = async (req, res) => {
         if (req.user.role !== 'requester') {
             return res.status(403).json({ status: 'fail', message: 'you are not allowed to update a trip request' })
         }
-        const requestId = req.params.id
+        const requestId = req.params.id;
+        const userId = req.user.id;
+        console.log('user', userId);
+        console.log('request', requestId);
         const tripRequest = await tripRequests.findOne({
             where: { id: requestId }
         })
+
 
         if (!tripRequest) {
             return res.status(404).json({ status: 'fail', message: `No Trip Request with id = ${requestId}` })
         }
 
         if (tripRequest.status == 'pending') {
-            const type = req.body.returnDate == ' ' ? 'one way trip' : 'return type trip';
+            const type = req.body.returnDate == null ? 'one way trip' : 'return type trip';
             const updatedTrip = {
                 leavingFrom: req.body.leavingFrom,
                 goingTo: req.body.goingTo,
@@ -122,7 +125,7 @@ export const updateTripRequest = async (req, res) => {
                 accomodationId: req.body.accomodationId
             }
 
-            tripRequest.update(updatedTrip, { where: { id: requestId } })
+            tripRequests.update(updatedTrip, { where: { id: requestId, requesterId: userId } })
                 .then(num => {
                     if (num == 1) {
                         res.send({ status: 'success', message: 'trip request updated successfully.' })
@@ -145,9 +148,11 @@ export const updateTripRequest = async (req, res) => {
 export const deleteTripRequest = async (req, res) => {
     try {
         if (req.user.role !== 'requester') {
-            return res.status(403).json({ status: 'fail', message: 'you are not allowed to update a trip request' })
+            return res.status(403).json({ status: 'fail', message: 'you are not allowed to delete a trip request' })
         }
-        const requestId = req.params.id
+        const requestId = req.params.id;
+        const userId = req.user.id;
+
         const tripRequest = await tripRequests.findOne({
             where: { id: requestId }
         })
@@ -157,11 +162,10 @@ export const deleteTripRequest = async (req, res) => {
         }
 
         if (tripRequest.status == 'pending') {
-            await tripRequests.destroy({ where: { id: requestId } })
+            await tripRequests.destroy({ where: { id: requestId, requesterId: userId } })
             res.status(200).json({ status: 'success', message: 'Trip Request Deleted successfully' })
         } else {
             return res.status(404).json({ status: 'fail', message: `Trip Request with id = ${requestId} is approved or rejected` })
-
         }
 
     } catch (error) {
